papertrade-backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg[binary]>=3.1.10
python-decouple==3.8
dj-database-url==2.1.0
PyJWT==2.8.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
pandas>=2.1.0
numpy>=1.26.0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
faker==20.1.0
